facial-recognition/requirements.txt
flask==2.3.3
dlib==19.24.0
gunicorn==21.2.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0